split util.py
