+++ .travis.yml
@@ -32,7 +32,7 @@ services:
 - xvfb
 before_install:
 - openssl aes-256-cbc -K $encrypted_0fabcfc21ca1_key -iv $encrypted_0fabcfc21ca1_iv
-  -in ./mmeds/secrets.py.enc -out mmeds/secrets.py -d
+  -in ./mmeds/secrets.py.enc -out ./mmeds/secrets.py -d
 - openssl aes-256-cbc -K $encrypted_754fbe69cba7_key -iv $encrypted_754fbe69cba7_iv
   -in ./mmeds/resources/privkey.pem.enc -out ./mmeds/resources/privkey.pem -d
 - wget http://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh -O ~/miniconda.sh
