+++ mmeds/test_files/test_metadata.tsv
@@ -6,8 +6,8 @@ No Limit	No Limit	Limit 45 Characters	Kilogram	Limit 45 Characters	Limit 45 Char
 0	Yes	here1	0.025	Post-anal gut	UBERON:3010437	Mountain	Mount Sinai	Charles	33234	66346	Grassland	Earth	-32990	122	Diamond	White	Illuminater	Lucy	63	2018-01-05	A19.8XXX	Good	2018-12-01	4	2017-02-21	Helpful	2018-02-22	2018-02-21	Heart Transplant	200	Surgery	Some.guy@gmail.com	Bill Hader	Amy Poehler	AGCTGACTAGTC	1_Fece_10_28_2008	GTGCCAGCMGCCGCGGTAA	alright1	L1S8	2018-01-21	Barack Obama	Protocol1	Full	Finished	12	E. Coli	Normal	Illumina	16S	Basic1	sc/orga/projects/	2018-03-21	Barack Obama	Results1	Protocol2	Qiime	2	Things1	2015-01-12	Barack Obama	Protocol1	Nothing	Normal	Illumina	0.5.1	2018-01-21	12:00	L1S8	Alot	0.0001	https://github.com/clementelab/mmeds-meta	Good_Study	Nice	1905	1	American	F	feces	UBERON:0001988	124	2018-01-05
 0	Yes	here2	0.025	Post-anal gut	UBERON:3010437	Mountain	Mount Sinai	Charles	33235	66347	Grassland	Earth	-32989	123	Diamond	White	Illuminater	Lucy	63	2018-01-06	A19.8XXX	Bad	2018-12-02	5	2017-02-22	Helpful	2018-02-23	2018-02-22	Heart Transplant	200	Surgery	Some.guy@gmail.com	Bill Hader	Amy Poehler	ATGGCAGCTCTA	2_Fece_10_28_2008	GTGCCAGCMGCCGCGGTAA	alright2	L1S140	2018-01-20	Barack Obama	Protocol2	Full	Finished	12	E. Coli	Normal	Illumina	16S	Basic2	sc/orga/projects/	2018-03-20	Barack Obama	Results2	Protocol2	Qiime	2	Things2	2015-01-12	Barack Obama	Protocol2	Nothing	Normal	Illumina	0.5.2	2018-01-20	12:01	L1S140	Alot	0.0001	https://github.com/clementelab/mmeds-meta	Good_Study	Nice	1906	2	American	F	feces	UBERON:0001988	125	2018-01-06
 84	No	here3	0.025	Post-anal gut	UBERON:3010437	Mountain	Mount Sinai	Charles	33236	66348	Grassland	Earth	-32988	124	Diamond	White	Illuminater	Lucy	63	2018-01-07	A19.8XXX	Ugly	2018-12-03	6	2017-02-23	Helpful	2018-02-24	2018-02-23	Heart Transplant	200	Surgery	Some.guy@gmail.com	Bill Hader	Amy Poehler	ACACACTATGGC	1_Fece_1_20_2009	GTGCCAGCMGCCGCGGTAA	alright3	L1S57	2018-01-19	Barack Obama	Protocol3	Full	Finished	12	E. Coli	Normal	Illumina	16S	Basic3	sc/orga/projects/	2018-03-19	Barack Obama	Results3	Protocol2	Qiime	2	Things3	2015-01-12	Barack Obama	Protocol3	Nothing	Normal	Illumina	0.5.3	2018-01-19	12:02	L1S57	Alot	0.0001	https://github.com/clementelab/mmeds-meta	Good_Study	Nice	1907	3	American	F	feces	UBERON:0001988	126	2018-01-07
-84	No	here4	0.025	Post-anal gut	UBERON:3010437	Mountain	Mount Sinai	Charles	33237	66349	Grassland	Earth	-32987	125	Diamond	White	Illuminater	Lucy	63	2018-01-08	A19.8XXX	Bad	2018-12-04	7	2017-02-24	Helpful	2018-02-25	2018-02-24	Heart Transplant	200	Surgery	Some.guy@gmail.com	Bill Hader	Amy Poehler	CTGAGATACGCG	2_Fece_1_20_2009	GTGCCAGCMGCCGCGGTAA	alright4	L1S208	2018-01-18	Barack Obama	Protocol4	Full	Finished	12	E. Coli	Normal	Illumina	16S	Basic4	sc/orga/projects/	2018-03-18	Barack Obama	Results4	Protocol3	Qiime	2	Things4	2015-01-13	Barack Obama	Protocol4	Nothing	Normal	Illumina	0.5.4	2018-01-18	12:03	L1S208	Alot	0.0001	https://github.com/clementelab/mmeds-meta	Good_Study	Nice	1908	4	American	F	feces	UBERON:0001988	127	2018-01-08
-112	No	here5	0.025	Post-anal gut	UBERON:3010437	Mountain	Mount Sinai	Charles	33238	66350	Grassland	Earth	-32986	126	Diamond	White	Illuminater	Lucy	63	2018-01-09	A19.8XXX	Bad	2018-12-05	8	2017-02-25	Helpful	2018-02-26	2018-02-25	Heart Transplant	200	Surgery	Some.guy@gmail.com	Bill Hader	Amy Poehler	ACTACGTGTGGT	1_Fece_2_17_2009	GTGCCAGCMGCCGCGGTAA	alright5	L1S76	2018-01-17	Barack Obama	Protocol5	Full	Finished	12	E. Coli	Normal	Illumina	16S	Basic5	sc/orga/projects/	2018-03-17	Barack Obama	Results5	Protocol3	Qiime	2	Things5	2015-01-13	Barack Obama	Protocol5	Nothing	Normal	Illumina	0.5.5	2018-01-17	12:04	L1S76	Alot	0.0001	https://github.com/clementelab/mmeds-meta	Good_Study	Nice	1909	5	American	F	feces	UBERON:0001988	128	2018-01-09
+84	No	here4	0.025	Post-anal gut	UBERON:3010437	Mountain	Mount Sinai	Charles	33237	66349	Grassland	Earth	-32987	125	Diamond	White	Illuminater	Lucy	63	2018-01-08	NA	Bad	2018-12-04	7	2017-02-24	Helpful	2018-02-25	2018-02-24	Heart Transplant	200	Surgery	Some.guy@gmail.com	Bill Hader	Amy Poehler	CTGAGATACGCG	2_Fece_1_20_2009	GTGCCAGCMGCCGCGGTAA	alright4	L1S208	2018-01-18	Barack Obama	Protocol4	Full	Finished	12	E. Coli	Normal	Illumina	16S	Basic4	sc/orga/projects/	2018-03-18	Barack Obama	Results4	Protocol3	Qiime	2	Things4	2015-01-13	Barack Obama	Protocol4	Nothing	Normal	Illumina	0.5.4	2018-01-18	12:03	L1S208	Alot	0.0001	https://github.com/clementelab/mmeds-meta	Good_Study	Nice	1908	4	American	F	feces	UBERON:0001988	127	2018-01-08
+112	No	here5	0.025	Post-anal gut	UBERON:3010437	Mountain	Mount Sinai	Charles	33238	66350	Grassland	Earth	-32986	126	Diamond	White	Illuminater	Lucy	63	2018-01-09	NA	Bad	2018-12-05	8	2017-02-25	Helpful	2018-02-26	2018-02-25	Heart Transplant	200	Surgery	Some.guy@gmail.com	Bill Hader	Amy Poehler	ACTACGTGTGGT	1_Fece_2_17_2009	GTGCCAGCMGCCGCGGTAA	alright5	L1S76	2018-01-17	Barack Obama	Protocol5	Full	Finished	12	E. Coli	Normal	Illumina	16S	Basic5	sc/orga/projects/	2018-03-17	Barack Obama	Results5	Protocol3	Qiime	2	Things5	2015-01-13	Barack Obama	Protocol5	Nothing	Normal	Illumina	0.5.5	2018-01-17	12:04	L1S76	Alot	0.0001	https://github.com/clementelab/mmeds-meta	Good_Study	Nice	1909	5	American	F	feces	UBERON:0001988	128	2018-01-09
 140	No	here6	0.025	Post-anal gut	UBERON:3010437	Mountain	Mount Sinai	Charles	33239	66351	Grassland	Earth	-32985	127	Diamond	White	Illuminater	Lucy	63	2018-01-10	A34.XXXX	Bad	2018-12-06	9	2017-02-26	Helpful	2018-02-27	2018-02-26	Heart Transplant	200	Surgery	Some.guy@gmail.com	Bill Hader	Amy Poehler	AGTGCGATGCGT	1_Fece_3_17_2009	GTGCCAGCMGCCGCGGTAA	alright6	L1S105	2018-01-16	Barack Obama	Protocol6	Full	Finished	12	E. Coli	Normal	Illumina	16S	Basic6	sc/orga/projects/	2018-03-16	Barack Obama	Results6	Protocol3	Qiime	2	Things6	2015-01-13	Barack Obama	Protocol6	Nothing	Normal	Illumina	0.5.6	2018-01-16	12:05	L1S105	Alot	0.0001	https://github.com/clementelab/mmeds-meta	Good_Study	Nice	1910	6	American	F	feces	UBERON:0001988	129	2018-01-10
 140	No	here7	0.025	Post-anal gut	UBERON:3010437	Mountain	Mount Sinai	Charles	33240	66352	Grassland	Earth	-32984	128	Diamond	White	Illuminater	Lucy	63	2018-01-11	A34.XXXX	Bad	2018-12-07	10	2017-02-27	Helpful	2018-02-28	2018-02-27	Heart Transplant	200	Surgery	Some.guy@gmail.com	Bill Hader	Amy Poehler	CCGACTGAGATG	2_Fece_3_17_2009	GTGCCAGCMGCCGCGGTAA	alright7	L1S257	2018-01-15	Barack Obama	Protocol7	Full	Finished	12	E. Coli	Normal	Illumina	16S	Basic7	sc/orga/projects/	2018-03-15	Barack Obama	Results7	Protocol4	Qiime	2	Things7	2015-01-14	Barack Obama	Protocol7	Nothing	Normal	Illumina	0.5.7	2018-01-15	12:06	L1S257	Alot	0.0001	https://github.com/clementelab/mmeds-meta	Good_Study	Nice	1911	7	American	F	feces	UBERON:0001988	130	2018-01-11
 168	No	here8	0.025	Post-anal gut	UBERON:3010437	Mountain	Mount Sinai	Charles	33241	66353	Grassland	Earth	-32983	129	Diamond	White	Illuminater	Lucy	63	2018-01-12	A34.XXXX	Bad	2018-12-08	11	2017-02-28	Helpful	2018-03-01	2018-02-28	Heart Transplant	200	Surgery	Some.guy@gmail.com	Bill Hader	Amy Poehler	CCTCTCGTGATC	2_Fece_4_14_2009	GTGCCAGCMGCCGCGGTAA	alright8	L1S281	2018-01-14	Barack Obama	Protocol8	Full	Finished	12	E. Coli	Normal	Illumina	16S	Basic8	sc/orga/projects/	2018-03-14	Barack Obama	Results8	Protocol4	Qiime	2	Things8	2015-01-14	Barack Obama	Protocol8	Nothing	Normal	Illumina	0.5.8	2018-01-14	12:07	L1S281	Alot	0.0001	https://github.com/clementelab/mmeds-meta	Good_Study	Nice	1912	8	American	F	feces	UBERON:0001988	131	2018-01-12
